refactor forms to put form components in shared. CSS tweaks to forms and modal. Add border-box to global CSS reset.
