+++ client/source/AppHolder.jsx
@@ -10,7 +10,7 @@ import * as requests from './utilities/axiosRequests.js';
 // MAIN
 const AppHolder = function CreateAppHolder() {
   // STATES
-  const [prodID, setProdID] = useState(40356); // TODO: default view for 'no item searched yet'
+  const [prodID, setProdID] = useState(40355); // TODO: default view for 'no item searched yet'
   const [product, setProduct] = useState(undefined);
   const [styles, setStyles] = useState([]);
   const [meta, setMeta] = useState({});
