+++ client/source/components/shared/FormTextarea.jsx
@@ -1,21 +1,21 @@
 import React from 'react';
 import styled from 'styled-components';
 
-const FormInput = function CreateFormInput({
-  name, value, labelText, placeholder, onChange, disclaimerText, maxLength = '60', type = 'text'
+const FormTextarea = function CreateFormTextarea({
+  labelText, name, value, placeholder, onChange, disclaimerText, maxLength = '1000'
 }) {
   return (
     <>
       <label htmlFor={name}>{labelText}
-        <Input
-          type={type}
-          name={name}
+        <Textarea
           id={name}
+          name={name}
           value={value}
-          maxLength={maxLength}
           placeholder={placeholder}
+          maxLength={maxLength}
           onChange={onChange}
-          cols="25"
+          rows="5"
+          cols="62"
         />
       </label>
       {disclaimerText && (<Paragraph>{disclaimerText}</Paragraph>)}
@@ -23,12 +23,13 @@ const FormInput = function CreateFormInput({
   );
 };
 
-const Input = styled.input`
-  width: 300px;
+const Textarea = styled.textarea`
+  font-family: inherit;
+  width: 100%;
 `;
 
 const Paragraph = styled.p`
   font-size: 75%;
 `;
 
-export default FormInput;
+export default FormTextarea;
